requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Auth
pyjwt==2.9.0
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
phonenumbers==8.13.47

# Utils
python-dotenv==1.0.1
pyotp==2.9.0

# HTTP Client (for adapters)
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Image processing (optional for photo validation)
pillow==10.4.0

# Геолокация
geopy==2.4.1
